Separated DB managing functions to db_manager.rb
