+++ lib/rubySQL/create.rb
@@ -1,5 +1,6 @@
 require 'sqlite3'
 require_relative 'assert'
+require_relative 'db_manager'
 
 class RubySQL::Create
   # Initializes a new class member variable with the passed db handler.
@@ -8,8 +9,9 @@ class RubySQL::Create
   # - db_name (str): Database name
   # Returns:
   # - None
-  def initialize(dbh, db_name)
+  def initialize(dbh, dbm, db_name)
     @dbh = dbh
+    @dbm = dbm
     @db_name = db_name
   end
 
@@ -29,7 +31,8 @@ class RubySQL::Create
 
     # Retrieve only the column names
     col_names = columns[0].keys
-    
+
+    #    
     table_spec_str = '('
     # col: Column name
     # columns[0][col][0]: Column type
@@ -59,7 +62,7 @@ class RubySQL::Create
   # Returns:
   # - None 
   def sqlite3_schema(table_name)
-    table_schema = sqlite3_pragma(table_name)
+    table_schema = @dbm.sqlite3_pragma(table_name)
 
     # First, find the max lengths of each column's title strings
     # for nice print out in a fixed length of table format
@@ -102,22 +105,13 @@ class RubySQL::Create
     }
   end
 
-  # Get and return the table PRAGMA
-  # Params:
-  # - table_name (str): Table name
-  # Returns:
-  # - None
-  def sqlite3_pragma(table_name)
-    return  @dbh.execute("PRAGMA table_info(#{table_name});")
-  end
-
   # Prints out the list of all tables in the database with their schema.
   # Params:
   # - None
   # Returns:
   # - None
   def sqlite3_list_tables
-    tables = sqlite3_all_tables
+    tables = @dbm.sqlite3_all_tables
     if !tables.empty?
       tables.each {|table|
         printf "#{table[1]} | #{table[4]}\n"
@@ -125,14 +119,5 @@ class RubySQL::Create
     else
       printf "Database [#{@db_name}] is empty.\n"      
     end
-  end
-
-  # Get and return the list of tables in the database
-  # Params:
-  # - None
-  # Returns:
-  # - tables (array): An array of tables.
-  def sqlite3_all_tables
-    return @dbh.execute("select * from sqlite_master where type='table';")
   end
 end
